Clean up webpack conf
Also makes it consistent with changes that are happening in the karma-based starters.

Connected to rangle/rangle-starter#127
